+++ docs/utils/inlinesyntaxhighlight.py
@@ -40,7 +40,7 @@ DIV_PRE_RE = re.compile(r'^<div[^>]*><pre>')
 PRE_DIV_RE = re.compile(r'\s*</pre></div>\s*$')
 
 def html_visit_literal(self, node):
-    env = self.settings.env
+    env = self.builder.env
 
     shall_highlight = False
 
@@ -57,7 +57,7 @@ def html_visit_literal(self, node):
 
         highlight_args = node.get('highlight_args', {})
 
-        if node.has_key('language'):
+        if 'language' in node.attributes:
             # code-block directives
             lang = node['language']
             highlight_args['force'] = True
@@ -146,7 +146,7 @@ def latex_visit_literal(self, node):
 
         highlight_args = node.get('highlight_args', {})
 
-        if node.has_key('language'):
+        if 'language' in node.attributes:
             # code-block directives
             lang = node['language']
             highlight_args['force'] = True